requirements.txt
Flask==2.3.3
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
flask-paginate==0.8.1
werkzeug==2.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==1.26.18